backend/requirements.txt
asgiref==3.11.1
attrs==26.1.0
dj-database-url==3.1.1
Django==6.0.4
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
gunicorn==23.0.0
iyzipay==1.0.45
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
psycopg==3.3.3
psycopg-binary==3.3.3
PyJWT==2.12.1
python-dotenv==1.2.2
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.5
stripe==15.0.1
tzdata==2026.2
whitenoise==6.11.0
uritemplate==4.2.0

